updated amount from double to BigDecimal

+++ src/main/java/com/revoltcode/account/common/event/account/AccountOpenedEvent.java
@@ -7,6 +7,7 @@ import lombok.Data;
 import lombok.NoArgsConstructor;
 import lombok.experimental.SuperBuilder;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 @Data
@@ -18,5 +19,5 @@ public class AccountOpenedEvent extends BaseEvent {
     private String name;
     private AccountType accountType;
     private LocalDateTime createdDate;
-    private double openingBalance;
+    private BigDecimal openingBalance;
 }
